[BUGFIX] Pass on translation elements via translate button
For the page module, the elements that may be translated are selected
both when rendering the page view and when rendering the element list in
the translation wizard. The former may easily be overwritten by a custom
page module renderer, as e.g. done in Flux/Fluidcontent, but for the
latter, this is harder, as only the page id, colPos and language id are
passed to the AJAX module.

When translating items in a nested container structure, the exact list
of items therefore needs to be passed back from the localization modal
to the translation controller (within the AJAX request), so the correct
elements are selected. This also requires overriding
AjaxLocalizationController in the respective extension.

Resolves: #20834 (sitegeist)

(cherry picked from commit a28dda4)

(cherry picked from commit 3a82c0f)

(cherry picked from commit 42f7989)

(cherry picked from commit 2b27986)

(cherry picked from commit fd054a5)